third_party/blink/web_tests/svg/hittest/singular-transform-4.html

<!DOCTYPE html>
    <title>SVG hover transform image test</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <style>
    body { margin: 0; }
    .p .a {
      transform: scaleX(0);
    }
    .p:hover .a {
      transform: scaleX(1);
      fill: red;
    }
    </style>
    <svg width="400" height="100">
      <g class="p">
        <image class="a" width="200" height="100" preserveAspectRatio="none"
               xlink:href="../custom/resources/red-checker.png"></image>
        <rect class="b" width="100" height="100" fill="green"/>
      </g>
    </svg>
    <script>
    test(() => {
      // eventSender must be available to simulate hover.
      assert_true(
        'eventSender' in window,
        'eventSender should exist for hover simulation'
      );
      eventSender.mouseMoveTo(150, 50);

      // Verify that .p:hover .a is not selectable for <image> element.
      assert_equals(
        document.querySelector('.p:hover .a'),
        null,
        "document.querySelector('.p:hover .a') should be null"
      );
    }, "SVG hover transform image test: <image> element remains non-selectable under .p:hover");
    </script>
